Commit 5
Fixed basket movement bug, honestly I am not entirely sure what the issue was; I rewrote the code two times and restarted visual studio/unity, and then it worked first try.

The error given was alon gthe lined of"object refference is not set to an instance of an obkect", so I knew it had to do with the basked not being instanced properly.